lucene 7.2.1 , bcel 6.2

+++ src/org/opensolaris/opengrok/web/SearchHelper.java
@@ -167,7 +167,7 @@ public class SearchHelper {
     /**
      * total number of hits
      */
-    public int totalHits;
+    public long totalHits;
     /**
      * the query created by the used {@link QueryBuilder} via
      * {@link #prepareExec(SortedSet)}.
